requirements.txt
Flask==0.11.1
flask-restplus==0.9.2
httplib2==0.9.2
PyJWT==1.4.2
pymongo==3.3.0